+++ vcf/src/vcf/ApplicationKit/Dialog.cpp
@@ -282,7 +282,7 @@ void Dialog::showMessage( const String& message, const String& caption )
 }
 
 UIToolkit::ModalReturnType Dialog::showMessage( const String& message, const String& caption,
-														const long& messageButtons,
+														const int32& messageButtons,
 														const MessageStyle& messageStyle )
 {
 	UIToolkit::ModalReturnType result = UIToolkit::mrOK;
